validattion added for equal ballots count for entering in representative's mode, bug fixed when deleteing party members list, candidates, who are in single mandates county only will not be deleted

+++ src/main/java/lt/itakademija/electors/candidate/CandidateService.java
@@ -53,8 +53,9 @@ public class CandidateService {
         List<CandidateReport> candidates = partyService.getPartyById(id).getMembers();
         partyService.detach(partyEntity);
         for (CandidateReport candidate : candidates) {
-            System.out.println(candidate.getId());
-            repository.delete(candidate.getId());
+            if (candidate.isMultiList() == true) {
+                repository.delete(candidate.getId());
+            }
         }
     return true;
     }
